Realign codebase to new termination API
We now distinguish between `CanTerminate()` and `OnCommittedState()`,
which were previously joined into the single `OnGVT()` API. This is in
preparation for the new termination detection.

Simulation models have been realigned accordingly.

Incidentally, the tests scripts has been updated to delete previous run
binary, to avoid re-running twice the same test, while thinking that a
different test was successful. This was not a problem on Travis, but
could be a problem locally.

+++ models/pcs/functions_app.c
@@ -170,7 +170,7 @@ int allocation(lp_state_type *pointer) {
 			}
 		}
 
-		if (fabsf(summ) < FLT_EPSILON) {
+		if (fabs(summ) < FLT_EPSILON) {
 			// The newly allocated channel receives the minimal power
 			c->sir_data->power = MIN_POWER;
 		} else {
